+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * build tools update
 * add SDK name
 * German translation from https://github.com/c-4-m
+* Excessive retrigger counter for concurrent audio loss via UI focus
 
 ### 4.5.2 (2020-10-16)
 * bugfix for adding already added fragment